Extract javaMail as separate plugin
