+++ appserver/admingui/core/src/main/resources/org/glassfish/admingui/core/Strings.properties
@@ -64,8 +64,8 @@ tree.connectionFactories=Connection Factories
 tree.connectionFactories.tooltip=Connection Factories
 tree.destinationResources=Destination Resources
 tree.destinationResources.tooltip=Destination Resources
-tree.javaMailSessions=JavaMail Sessions
-tree.javaMailSessions.tooltip=JavaMail Sessions
+tree.javaMailSessions=Jakarta Mail Sessions
+tree.javaMailSessions.tooltip=Jakarta Mail Sessions
 tree.jndi=JNDI
 tree.jndi.tooltip=JNDI
 tree.customResources=Custom Resources
