+++ appserver/admingui/full/src/main/help/en/help/idx-1.html
@@ -51,17 +51,6 @@
 <dd>Java Naming and Directory Interface, <i>See </i>JNDI</dd>
 </dl>
 <dl>
-<dd>JavaMail sessions<dl>
-<dd>creating (reference), <a href="ref-javamailsessioncreate.html#indexterm-22">New JavaMail Session</a></dd>
-<dd>creating (task), <a href="task-javamailsessioncreate.html#indexterm-1">To Create a JavaMail Session</a></dd>
-<dd>deleting, <a href="task-javamailsessionsdelete.html#indexterm-3">To Delete a JavaMail Session</a></dd>
-<dd>editing (reference), <a href="ref-javamailsessionsedit.html#indexterm-23">Edit JavaMail Session</a></dd>
-<dd>editing (task), <a href="task-javamailsessionsedit.html#indexterm-2">To Edit a JavaMail Session</a></dd>
-<dd>viewing and configuring, <a href="ref-javamailsessions.html#indexterm-21">JavaMail Sessions</a></dd>
-</dl>
-</dd>
-</dl>
-<dl>
 <dd>JNDI custom resources<dl>
 <dd>creating (reference), <a href="ref-jndicustomresourcenew.html#indexterm-28">New Custom Resource</a></dd>
 <dd>creating (task), <a href="task-jndicustomresourcenew.html#indexterm-6">To Create a Custom JNDI Resource</a></dd>
