+++ appserver/admingui/full/src/main/help/en/help/index.xml
@@ -39,14 +39,6 @@
     <indexitem text="editing (task)" target="task-jndiexternalresourceedit"></indexitem>
     <indexitem text="viewing and configuring" target="ref-jndiexternalresources"></indexitem>
   </indexitem>
-  <indexitem text="JavaMail sessions">
-    <indexitem text="creating (reference)" target="ref-javamailsessioncreate"></indexitem>
-    <indexitem text="creating (task)" target="task-javamailsessioncreate"></indexitem>
-    <indexitem text="deleting" target="task-javamailsessionsdelete"></indexitem>
-    <indexitem text="editing (reference)" target="ref-javamailsessionsedit"></indexitem>
-    <indexitem text="editing (task)" target="task-javamailsessionsedit"></indexitem>
-    <indexitem text="viewing and configuring" target="ref-javamailsessions"></indexitem>
-  </indexitem>
   <indexitem text="JNDI custom resources">
     <indexitem text="creating (reference)" target="ref-jndicustomresourcenew"></indexitem>
     <indexitem text="creating (task)" target="task-jndicustomresourcenew"></indexitem>
