+++ appserver/admingui/full/src/main/resources/org/glassfish/full/admingui/Helplinks.properties
@@ -37,15 +37,6 @@ externalResourceEdit=ref-jndiexternalresourceedit.html
 externalResourceTargets=ref-jndiexternalresourcetargets.html
 
 ##############################
-# JAVAMAIL SESSIONS
-##############################
-javaMailSessions=ref-javamailsessions.html
-javaMailSessionNew=ref-javamailsessioncreate.html
-javaMailSessionEdit=ref-javamailsessionsedit.html
-javaMailSessionTargets=ref-javamailsessiontargets.html
-
-
-##############################
 # Batch support
 ##############################
 batchJobs=ref-batchjobslisting.html
