+++ appserver/admingui/full/src/main/resources/org/glassfish/full/admingui/Strings.properties
@@ -14,35 +14,6 @@
 # SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 #
 
-## Java Mail Sessions
-javaMailSessions.pageTitle=JavaMail Sessions
-javaMailSessions.pageTitleHelp=A JavaMail session resource represents a mail session in the JavaMail API, which provides a platform-independent and protocol-independent framework to build mail and messaging applications.
-javaMail.editPageTitle=Edit JavaMail Session
-javaMail.editPageTitleHelp=A JavaMail session resource represents a mail session in the JavaMail API.
-javaMail.newPageTitle=New JavaMail Session
-javaMail.newPageTitleHelp=A JavaMail session resource represents a mail session in the JavaMail API.
-javaMail.targetPageTitle=JavaMail Session Targets
-javaMail.jndiHelp=A unique identifier; must contain only alphanumeric, underscore, dash, or dot characters
-javaMail.mailHost=Mail Host:
-javaMail.mailHostHelp=DNS name of the default mail server
-javaMail.user=Default User:
-javaMail.userHelp=User name to provide when connecting to a mail server; must contain only alphanumeric, underscore, dash, or dot characters
-javaMail.returnAddr=Default Sender Address:
-javaMail.returnAddrHelp=E-mail address of the default user
-javaMail.descHelp=Makes it easier to find this session later
-javaMail.Session=Session:
-javaMail.storeProtocol=Store Protocol:
-javaMail.storeProtocolHelp=Either IMAP or POP3; default is IMAP
-javaMail.storeClass=Store Protocol Class:
-javaMail.storeClassHelp=Default is com.sun.mail.imap.IMAPStore
-javaMail.transportProtocol=Transport Protocol:
-javaMail.transportProtocolHelp=Default is SMTP
-javaMail.transProtocolClass=Transport Protocol Class:
-javaMail.transProtocolClassHelp=Default is com.sun.mail.smtp.SMTPTransport
-javaMail.Debug=Debug:
-javaMail.DebugHelp=Select the Debug checkbox to enable extra debugging output for this mail session, including a protocol trace.
-
-
 ## Custom and JNDI Resource
 customResources.pageTitle=Custom Resources
 customResources.pageTitleHelp=Custom resources are nonstandard resources with a defined JNDI subcontext, resource type, and factory class.
