+++ appserver/admingui/mail-plugin/src/main/help/en/help/aseeacohjmail.html
@@ -16,7 +16,7 @@
 
 -->
 
-<p><a id="ref-group-javamailresources" name="ref-group-javamailresources"></a><a id="GHJML116" name="GHJML116"></a></p>
+<h1>GlassFish Server 3.1 Administration Console Jakarta Mail Online Help</h1>
 
-<h4>JavaMail Sessions</h4>
+
 <small>Copyright &#169; 2005, 2017, Oracle and/or its affiliates. All rights reserved. <a href="docinfo.html">Legal Notices</a></small>
