+++ appserver/admingui/mail-plugin/src/main/help/en/help/docinfo.html
@@ -1,0 +1,38 @@
+<!--
+
+    Copyright (c) 2005, 2019 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<h4>GlassFish Server Open Source Edition 7</h4>
+<p><a name="BEGIN" id="BEGIN"></a></p>
+<p>Jakarta Mail Online Help</p>
+<p></p>
+
+<p>August 2017</p>
+<p>This document provides the reference and task online help topics for the full module of the GlassFish Server Administration Console.</p>
+<p><a id="sthref1" name="sthref1"></a></p>
+<hr />
+<p>GlassFish Server Open Source Edition 7 Jakarta Mail Online Help</p>
+
+<p>Copyright &copy; 2010, 2019 Oracle and/or its affiliates. All rights reserved.</p>
+ 
+<p>This program and the accompanying materials are made available under the terms of the Eclipse Public License v. 2.0, which is available at  http://www.eclipse.org/legal/epl-2.0.</p> 
+ 
+<p>This Source Code may also be made available under the following Secondary Licenses when the conditions for such availability setforth in the Eclipse Public License v. 2.0 are satisfied: GNU General Public License, version 2 with the GNU Classpath Exception, which is available at https://www.gnu.org/software/classpath/license.html.</p>
+ 
+<p>SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0</p>
+
+
