+++ appserver/admingui/mail-plugin/src/main/help/en/help/idx-1.html
@@ -1,0 +1,34 @@
+<!--
+
+    Copyright (c) 2005, 2018 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<a name="idx"></a><h3>Index</h3><h2>J</h2>
+<div class="indexdiv">
+<dl>
+<dd>Jakarta Mail sessions<dl>
+<dd>creating (reference), <a href="ref-javamailsessioncreate.html#indexterm-22">New Jakarta Mail Session</a></dd>
+<dd>creating (task), <a href="task-javamailsessioncreate.html#indexterm-1">To Create a Jakarta Mail Session</a></dd>
+<dd>deleting, <a href="task-javamailsessionsdelete.html#indexterm-3">To Delete a Jakarta Mail Session</a></dd>
+<dd>editing (reference), <a href="ref-javamailsessionsedit.html#indexterm-23">Edit Jakarta Mail Session</a></dd>
+<dd>editing (task), <a href="task-javamailsessionsedit.html#indexterm-2">To Edit a Jakarta Mail Session</a></dd>
+<dd>viewing and configuring, <a href="ref-javamailsessions.html#indexterm-21">Jakarta Mail Sessions</a></dd>
+</dl>
+</dd>
+</dl>
+</div>
+
+<small>Copyright &#169; 2005, 2017, Oracle and/or its affiliates. All rights reserved. <a href="docinfo.html">Legal Notices</a></small>
