+++ appserver/admingui/mail-plugin/src/main/help/en/help/index.xml
@@ -1,0 +1,28 @@
+<!--
+
+    Copyright (c) 2017, 2018 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<index version="2.0">
+  <indexitem text="Jakarta Mail sessions">
+    <indexitem text="creating (reference)" target="ref-javamailsessioncreate"></indexitem>
+    <indexitem text="creating (task)" target="task-javamailsessioncreate"></indexitem>
+    <indexitem text="deleting" target="task-javamailsessionsdelete"></indexitem>
+    <indexitem text="editing (reference)" target="ref-javamailsessionsedit"></indexitem>
+    <indexitem text="editing (task)" target="task-javamailsessionsedit"></indexitem>
+    <indexitem text="viewing and configuring" target="ref-javamailsessions"></indexitem>
+  </indexitem>
+</index>
