+++ appserver/admingui/mail-plugin/src/main/help/en/help/ref-group-javamailresources.html
@@ -16,7 +16,7 @@
 
 -->
 
-<h1>GlassFish Server 3.1 Administration Console JavaMail Online Help</h1>
+<p><a id="ref-group-javamailresources" name="ref-group-javamailresources"></a><a id="GHJML116" name="GHJML116"></a></p>
 
-
+<h4>Jakarta Mail Sessions</h4>
 <small>Copyright &#169; 2005, 2017, Oracle and/or its affiliates. All rights reserved. <a href="docinfo.html">Legal Notices</a></small>
