+++ appserver/admingui/mail-plugin/src/main/help/en/help/ref-javamailsessioncreate.html
@@ -18,14 +18,14 @@
 
 <p><a id="ref-javamailsessioncreate" name="ref-javamailsessioncreate"></a><a id="GHJML00002" name="GHJML00002"></a></p>
 
-<h4><a id="sthref23" name="sthref23"></a>New JavaMail Session</h4>
+<h4><a id="sthref23" name="sthref23"></a>New Jakarta Mail Session</h4>
 <a name="BEGIN" id="BEGIN"></a>
-<p>Use the New JavaMail Session page to create a JavaMail session resource.</p>
-<p>The New JavaMail Session page contains the following options.</p>
+<p>Use the New Jakarta Mail Session page to create a Jakarta Mail session resource.</p>
+<p>The New Jakarta Mail Session page contains the following options.</p>
 <dl>
 <dt>JNDI Name</dt>
 <dd>
-<p>A unique name that identifies the mail session. Use the naming sub-context prefix <code>mail/</code> for JavaMail resources. For example: <code>mail/MySession</code>. The name must contain only alphanumeric, underscore, dash, or dot characters.</p>
+<p>A unique name that identifies the mail session. Use the naming sub-context prefix <code>mail/</code> for Jakarta Mail resources. For example: <code>mail/MySession</code>. The name must contain only alphanumeric, underscore, dash, or dot characters.</p>
 </dd>
 <dt>Mail Host</dt>
 <dd>
@@ -89,11 +89,11 @@
 </dd>
 <dt>Debug</dt>
 <dd>
-<p>If this option is selected, debugging for this resource is enabled. If the JavaMail log level is set to <code>FINE</code> or finer, the debugging output is generated and is included in the system log file. See <a href="../../../common/en/help/task-eelogginglevels.html">To Configure Log Levels</a> for information about setting the log level.</p>
+<p>If this option is selected, debugging for this resource is enabled. If the Jakarta Mail log level is set to <code>FINE</code> or finer, the debugging output is generated and is included in the system log file. See <a href="../../../common/en/help/task-eelogginglevels.html">To Configure Log Levels</a> for information about setting the log level.</p>
 </dd>
 <dt>Additional Properties</dt>
 <dd>
-<p>Additional properties for the JavaMail session. For a list of the available properties, see the JavaMail API documentation (<code><a href="http://java.sun.com/products/javamail/javadocs/index.html">http://java.sun.com/products/javamail/javadocs/index.html</a></code>). Additional properties must begin with <code>mail-</code> and must use hyphens, not periods, as separators. For example, <code>mail-debug</code> is correct, but <code>mail.debug</code> is not. The GlassFish Server back end converts the hyphens into the periods expected by the JavaMail API.</p>
+<p>Additional properties for the Jakarta Mail session. For a list of the available properties, see the Jakarta Mail API documentation (<code><a href="http://java.sun.com/products/javamail/javadocs/index.html">http://java.sun.com/products/javamail/javadocs/index.html</a></code>). Additional properties must begin with <code>mail-</code> and must use hyphens, not periods, as separators. For example, <code>mail-debug</code> is correct, but <code>mail.debug</code> is not. The GlassFish Server back end converts the hyphens into the periods expected by the Jakarta Mail API.</p>
 </dd>
 <dt>Targets</dt>
 <dd>
@@ -105,13 +105,13 @@
 <h5>Related Tasks</h5>
 <ul>
 <li>
-<p><a href="task-javamailsessioncreate.html">To Create a JavaMail Session</a></p>
+<p><a href="task-javamailsessioncreate.html">To Create a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsedit.html">To Edit a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsedit.html">To Edit a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsdelete.html">To Delete a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsdelete.html">To Delete a Jakarta Mail Session</a></p>
 </li>
 </ul>
 <a id="GHJML120" name="GHJML120"></a>
