+++ appserver/admingui/mail-plugin/src/main/help/en/help/ref-javamailsessions.html
@@ -18,15 +18,15 @@
 
 <p><a id="ref-javamailsessions" name="ref-javamailsessions"></a><a id="GHJML00001" name="GHJML00001"></a></p>
 
-<h4><a id="sthref22" name="sthref22"></a>JavaMail Sessions</h4>
+<h4><a id="sthref22" name="sthref22"></a>Jakarta Mail Sessions</h4>
 <a name="BEGIN" id="BEGIN"></a>
-<p>GlassFish Server includes the JavaMail API. JavaMail provides access from Java applications to Internet Message Access Protocol (IMAP) and Simple Mail Transfer Protocol (SMTP) capable mail servers on your network or the Internet.</p>
-<p>Use the JavaMail Sessions page to configure JavaMail sessions.</p>
-<p>For each JavaMail session, the following information is provided.</p>
+<p>GlassFish Server includes the Jakarta Mail API. Jakarta Mail provides access from Java applications to Internet Message Access Protocol (IMAP) and Simple Mail Transfer Protocol (SMTP) capable mail servers on your network or the Internet.</p>
+<p>Use the Jakarta Mail Sessions page to configure Jakarta Mail sessions.</p>
+<p>For each Jakarta Mail session, the following information is provided.</p>
 <dl>
 <dt>JNDI Name</dt>
 <dd>
-<p>A unique name that identifies the JavaMail session.</p>
+<p>A unique name that identifies the Jakarta Mail session.</p>
 </dd>
 <dt>Enabled</dt>
 <dd>
@@ -40,39 +40,39 @@
 </dd>
 <dt>Description</dt>
 <dd>
-<p>A description of the JavaMail session.</p>
+<p>A description of the Jakarta Mail session.</p>
 </dd>
 </dl>
 <p>The Sessions table also contains the following options.</p>
 <dl>
 <dt>New</dt>
 <dd>
-<p>Button to create a new JavaMail session.</p>
+<p>Button to create a new Jakarta Mail session.</p>
 </dd>
 <dt>Delete</dt>
 <dd>
-<p>Button to delete one or more selected JavaMail sessions.</p>
+<p>Button to delete one or more selected Jakarta Mail sessions.</p>
 </dd>
 <dt>Enable</dt>
 <dd>
-<p>Button to enable one or more selected JavaMail sessions.</p>
+<p>Button to enable one or more selected Jakarta Mail sessions.</p>
 </dd>
 <dt>Disable</dt>
 <dd>
-<p>Button to disable one or more selected JavaMail sessions.</p>
+<p>Button to disable one or more selected Jakarta Mail sessions.</p>
 </dd>
 </dl>
 <a id="GHJML117" name="GHJML117"></a>
 <h5>Related Tasks</h5>
 <ul>
 <li>
-<p><a href="task-javamailsessioncreate.html">To Create a JavaMail Session</a></p>
+<p><a href="task-javamailsessioncreate.html">To Create a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsedit.html">To Edit a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsedit.html">To Edit a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsdelete.html">To Delete a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsdelete.html">To Delete a Jakarta Mail Session</a></p>
 </li>
 </ul>
 <a id="GHJML118" name="GHJML118"></a>
