+++ appserver/admingui/mail-plugin/src/main/help/en/help/ref-javamailsessionsedit.html
@@ -18,10 +18,10 @@
 
 <p><a id="ref-javamailsessionsedit" name="ref-javamailsessionsedit"></a><a id="GHJML00003" name="GHJML00003"></a></p>
 
-<h4><a id="sthref24" name="sthref24"></a>Edit JavaMail Session</h4>
+<h4><a id="sthref24" name="sthref24"></a>Edit Jakarta Mail Session</h4>
 <a name="BEGIN" id="BEGIN"></a>
-<p>Use the Edit JavaMail Session page to edit a JavaMail session.</p>
-<p>The Edit JavaMail Session page contains the following options.</p>
+<p>Use the Edit Jakarta Mail Session page to edit a Jakarta Mail session.</p>
+<p>The Edit Jakarta Mail Session page contains the following options.</p>
 <dl>
 <dt>Load Defaults</dt>
 <dd>
@@ -29,7 +29,7 @@
 </dd>
 <dt>JNDI Name</dt>
 <dd>
-<p>A unique name that identifies the JavaMail session. The JNDI name is a read-only field. You can only specify the JNDI name when you create a new JavaMail session.</p>
+<p>A unique name that identifies the Jakarta Mail session. The JNDI name is a read-only field. You can only specify the JNDI name when you create a new Jakarta Mail session.</p>
 </dd>
 <dt>Mail Host</dt>
 <dd>
@@ -45,7 +45,7 @@
 </dd>
 <dt>Deployment Order</dt>
 <dd>
-<p>The deployment order of the JavaMail session resource.</p>
+<p>The deployment order of the Jakarta Mail session resource.</p>
 </dd>
 <dt>Description</dt>
 <dd>
@@ -99,24 +99,24 @@
 </dd>
 <dt>Debug</dt>
 <dd>
-<p>If this option is selected, debugging for this resource is enabled. If the JavaMail log level is set to <code>FINE</code> or finer, the debugging output is generated and is included in the system log file. See <a href="../../../common/en/help/task-eelogginglevels.html">To Configure Log Levels</a> for information about setting the log level.</p>
+<p>If this option is selected, debugging for this resource is enabled. If the Jakarta Mail log level is set to <code>FINE</code> or finer, the debugging output is generated and is included in the system log file. See <a href="../../../common/en/help/task-eelogginglevels.html">To Configure Log Levels</a> for information about setting the log level.</p>
 </dd>
 <dt>Additional Properties</dt>
 <dd>
-<p>Additional properties for the JavaMail session. For a list of the available properties, see the JavaMail API documentation (<code><a href="http://java.sun.com/products/javamail/javadocs/index.html">http://java.sun.com/products/javamail/javadocs/index.html</a></code>). Additional properties must begin with <code>mail-</code> and must use hyphens, not periods, as separators. For example, <code>mail-debug</code> is correct, but <code>mail.debug</code> is not. The GlassFish Server back end converts the hyphens into the periods expected by the JavaMail API.</p>
+<p>Additional properties for the Jakarta Mail session. For a list of the available properties, see the Jakarta Mail API documentation (<code><a href="http://java.sun.com/products/javamail/javadocs/index.html">http://java.sun.com/products/javamail/javadocs/index.html</a></code>). Additional properties must begin with <code>mail-</code> and must use hyphens, not periods, as separators. For example, <code>mail-debug</code> is correct, but <code>mail.debug</code> is not. The GlassFish Server back end converts the hyphens into the periods expected by the Jakarta Mail API.</p>
 </dd>
 </dl>
 <a id="GHJML121" name="GHJML121"></a>
 <h5>Related Tasks</h5>
 <ul>
 <li>
-<p><a href="task-javamailsessioncreate.html">To Create a JavaMail Session</a></p>
+<p><a href="task-javamailsessioncreate.html">To Create a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsedit.html">To Edit a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsedit.html">To Edit a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsdelete.html">To Delete a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsdelete.html">To Delete a Jakarta Mail Session</a></p>
 </li>
 </ul>
 <a id="GHJML122" name="GHJML122"></a>
