+++ appserver/admingui/mail-plugin/src/main/help/en/help/ref-topic-group.html
@@ -16,36 +16,6 @@
 
 -->
 
-<p><a id="task-javamailsessionsdelete" name="task-javamailsessionsdelete"></a><a id="GHJML00014" name="GHJML00014"></a></p>
-
-<h4><a id="sthref6" name="sthref6"></a>To Delete a JavaMail Session</h4>
-<a name="BEGIN" id="BEGIN"></a>
-<ol>
-<li>
-<p>In the navigation tree, expand the Resources node.</p>
-</li>
-<li>
-<p>Under the Resources node, select JavaMail Sessions.</p>
-<p>The JavaMail Sessions page opens.</p>
-</li>
-<li>
-<p>On the JavaMail Sessions page, select the checkbox next to the name of the session that you are deleting.</p>
-</li>
-<li>
-<p>Click Delete.</p>
-</li>
-</ol>
-<a id="GHJML106" name="GHJML106"></a>
-<h5>See Also</h5>
-<ul>
-<li>
-<p><a href="task-javamailsessioncreate.html">To Create a JavaMail Session</a></p>
-</li>
-<li>
-<p><a href="task-javamailsessionsedit.html">To Edit a JavaMail Session</a></p>
-</li>
-</ul>
-
-
-
+<p><a id="ref-topic-group" name="ref-topic-group"></a><a id="GHJML114" name="GHJML114"></a></p>
+<h4>GlassFish Server Administration Console Reference</h4>
 <small>Copyright &#169; 2005, 2017, Oracle and/or its affiliates. All rights reserved. <a href="docinfo.html">Legal Notices</a></small>
