+++ appserver/admingui/mail-plugin/src/main/help/en/help/task-group-javamailresources.html
@@ -18,5 +18,5 @@
 
 <p><a id="task-group-javamailresources" name="task-group-javamailresources"></a><a id="GHJML102" name="GHJML102"></a></p>
 
-<h4>Configuring JavaMail Sessions</h4>
+<h4>Configuring Jakarta Mail Sessions</h4>
 <small>Copyright &#169; 2005, 2017, Oracle and/or its affiliates. All rights reserved. <a href="docinfo.html">Legal Notices</a></small>
