+++ appserver/admingui/mail-plugin/src/main/help/en/help/task-javamailsessionsdelete.html
@@ -1,0 +1,51 @@
+<!--
+
+    Copyright (c) 2005, 2018 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<p><a id="task-javamailsessionsdelete" name="task-javamailsessionsdelete"></a><a id="GHJML00014" name="GHJML00014"></a></p>
+
+<h4><a id="sthref6" name="sthref6"></a>To Delete a Jakarta Mail Session</h4>
+<a name="BEGIN" id="BEGIN"></a>
+<ol>
+<li>
+<p>In the navigation tree, expand the Resources node.</p>
+</li>
+<li>
+<p>Under the Resources node, select Jakarta Mail Sessions.</p>
+<p>The Jakarta Mail Sessions page opens.</p>
+</li>
+<li>
+<p>On the Jakarta Mail Sessions page, select the checkbox next to the name of the session that you are deleting.</p>
+</li>
+<li>
+<p>Click Delete.</p>
+</li>
+</ol>
+<a id="GHJML106" name="GHJML106"></a>
+<h5>See Also</h5>
+<ul>
+<li>
+<p><a href="task-javamailsessioncreate.html">To Create a Jakarta Mail Session</a></p>
+</li>
+<li>
+<p><a href="task-javamailsessionsedit.html">To Edit a Jakarta Mail Session</a></p>
+</li>
+</ul>
+
+
+
+<small>Copyright &#169; 2005, 2017, Oracle and/or its affiliates. All rights reserved. <a href="docinfo.html">Legal Notices</a></small>
