+++ appserver/admingui/mail-plugin/src/main/help/en/help/task-javamailtargets.html
@@ -18,23 +18,23 @@
 
 <p><a id="task-javamailtargets" name="task-javamailtargets"></a><a id="GHJML00013" name="GHJML00013"></a></p>
 
-<h4>To Manage JavaMail Session Targets</h4>
+<h4>To Manage Jakarta Mail Session Targets</h4>
 <a name="BEGIN" id="BEGIN"></a>
 <ol>
 <li>
 <p>In the navigation tree, expand the Resources node.</p>
 </li>
 <li>
-<p>Under the Resources node, select JavaMail Sessions.</p>
-<p>The JavaMail Sessions page opens.</p>
+<p>Under the Resources node, select Jakarta Mail Sessions.</p>
+<p>The Jakarta Mail Sessions page opens.</p>
 </li>
 <li>
 <p>On the Sessions table, click the session name whose targets you want to manage.</p>
-<p>The Edit JavaMail Session page opens.</p>
+<p>The Edit Jakarta Mail Session page opens.</p>
 </li>
 <li>
 <p>Click the Target tab.</p>
-<p>The JavaMail Session Targets page opens.</p>
+<p>The Jakarta Mail Session Targets page opens.</p>
 </li>
 <li>
 <p>In the Targets table, choose the action you want to take:</p>
@@ -55,13 +55,13 @@
 <h5>See Also</h5>
 <ul>
 <li>
-<p><a href="task-javamailsessioncreate.html">To Create a JavaMail Session</a></p>
+<p><a href="task-javamailsessioncreate.html">To Create a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsedit.html">To Edit a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsedit.html">To Edit a Jakarta Mail Session</a></p>
 </li>
 <li>
-<p><a href="task-javamailsessionsdelete.html">To Delete a JavaMail Session</a></p>
+<p><a href="task-javamailsessionsdelete.html">To Delete a Jakarta Mail Session</a></p>
 </li>
 </ul>
