+++ appserver/admingui/mail-plugin/src/main/help/en/help/tasks-topic-group.html
@@ -1,6 +1,6 @@
 <!--
 
-    Copyright (c) 2010, 2018 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 2005, 2018 Oracle and/or its affiliates. All rights reserved.
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -16,11 +16,6 @@
 
 -->
 
-<sun:property>
-    <sun:hyperlink
-        toolTip="$resource{i18n.tree.jndi.tooltip}"
-        url="/full/jndi.jsf" >
-        <sun:image url="/resource/images/jndi.gif" />
-        <sun:staticText text="$resource{i18n.tree.jndi}" />
-    </sun:hyperlink>
-</sun:property>
+<p><a id="tasks-topic-group" name="tasks-topic-group"></a><a id="GHJML101" name="GHJML101"></a></p>
+<h4>GlassFish Server Administration Console Tasks</h4>
+<small>Copyright &#169; 2005, 2017, Oracle and/or its affiliates. All rights reserved. <a href="docinfo.html">Legal Notices</a></small>
