+++ appserver/admingui/mail-plugin/src/main/help/en/help/toc.xml
@@ -1,0 +1,37 @@
+<!--
+
+    Copyright (c) 2017, 2018 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<toc version="2.0">
+  <tocitem target="tasks-topic-group" text="GlassFish Server Administration Console Tasks" expand="false">
+    <tocitem target="task-group-javamailresources" text="Configuring Jakarta Mail Sessions" expand="false">
+      <tocitem target="task-javamailsessioncreate" text="To Create a Jakarta Mail Session" expand="false"></tocitem>
+      <tocitem target="task-javamailsessionsedit" text="To Edit a Jakarta Mail Session" expand="false"></tocitem>
+      <tocitem target="task-javamailtargets" text="To Manage Jakarta Mail Session Targets" expand="false"></tocitem>
+      <tocitem target="task-javamailsessionsdelete" text="To Delete a Jakarta Mail Session" expand="false"></tocitem>
+    </tocitem>
+  </tocitem>
+  <tocitem target="ref-topic-group" text="GlassFish Server Administration Console Reference" expand="false">
+    <tocitem target="ref-group-resources" text="Resources" expand="false">
+      <tocitem target="ref-group-javamailresources" text="Jakarta Mail Sessions" expand="false">
+        <tocitem target="ref-javamailsessions" text="Jakarta Mail Sessions" expand="false"></tocitem>
+        <tocitem target="ref-javamailsessioncreate" text="New Jakarta Mail Session" expand="false"></tocitem>
+        <tocitem target="ref-javamailsessionsedit" text="Edit Jakarta Mail Session" expand="false"></tocitem>
+      </tocitem>
+    </tocitem>
+  </tocitem>
+</toc>
