+++ appserver/admingui/mail-plugin/src/main/resources/META-INF/admingui/console-config.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2010, 2018 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<console-config id="jmail">
+    <integration-point
+        id="jakartaMailResourceLink"
+        parentId="propSheetSection"
+        type="org.glassfish.admingui:resources"
+        priority="100"
+        content="jakartaMailResourceLink.jsf"
+    />
+    <integration-point
+        id="jakartaMailNodes"
+        type="org.glassfish.admingui:navNode"
+        priority="500"
+        parentId="resources"
+        content="jakartaMailTreeNode.jsf"
+    />
+    <!-- add dropdown option for the resource dropdown of cluster/standalone instance resource table. -->
+    <integration-point
+            id="jakartaMailResourceDropDown"
+            type="org.glassfish.admingui:resourceTypeDropdown"
+            priority="500"
+            content="jmail/mailResourceNew.jsf|org.glassfish.jmail.admingui.Strings|tree.javaMailSessions"
+    />
+    <integration-point
+            id="jakartaMailFilter"
+            type="org.glassfish.admingui:resFilterDropdown"
+            priority="500"
+            content="mail-resource|org.glassfish.jmail.admingui.Strings|tree.javaMailSessions"
+    />
+</console-config>
