+++ appserver/admingui/mail-plugin/src/main/resources/jakartaMailResourceLink.jsf
@@ -16,11 +16,15 @@
 
 -->
 
+<!initPage
+    setResourceBundle(key="i18njmail" bundle="org.glassfish.jmail.admingui.Strings")
+/>
+
 <sun:property>
-    <sun:hyperlink
-        toolTip="$resource{i18n.tree.jndi.tooltip}"
-        url="/full/jndi.jsf" >
-        <sun:image url="/resource/images/jndi.gif" />
-        <sun:staticText text="$resource{i18n.tree.jndi}" />
+    <sun:hyperlink 
+        toolTip="$resource{i18njmail.tree.javaMailSessions.tooltip}"
+        url="/jmail/jakartaMailResources.jsf" >
+        <sun:image url="/resource/images/javaMailSession.gif" />
+        <sun:staticText text="$resource{i18njmail.tree.javaMailSessions}" />
     </sun:hyperlink>
 </sun:property>
