+++ appserver/admingui/mail-plugin/src/main/resources/mailResourceAttr.inc
