+++ appserver/admingui/mail-plugin/src/main/resources/mailResourceEdit.jsf
@@ -18,8 +18,8 @@
 
 <!-- full/jmailResourceEdit.jsf -->
 <!initPage
-    setResourceBundle(key="i18njmail" bundle="org.glassfish.full.admingui.Strings");
-    setResourceBundle(key="help_jmail" bundle="org.glassfish.full.admingui.Helplinks")
+    setResourceBundle(key="i18njmail" bundle="org.glassfish.jmail.admingui.Strings");
+    setResourceBundle(key="help_jmail" bundle="org.glassfish.jmail.admingui.Helplinks")
 />
 <!composition template="/templates/default.layout"  guiTitle="$resource{i18njmail.javaMail.editPageTitle}" >
 <!define name="content">
@@ -33,18 +33,18 @@
         getRequestValue(key="moduleName" value="#{pageSession.moduleName}");
         if(#{pageSession.moduleName}) {
             urlencode(value="#{pageSession.moduleName}" encoding="UTF-8" result="#{pageSession.encodedModuleName}");
-            setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/full/mailResourceEdit.jsf?name=#{pageSession.encodedName}&isAppScopedRes=$boolean{true}&appName=#{pageSession.encodedAppName}&moduleName=#{pageSession.encodedModuleName}");
+            setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/jmail/mailResourceEdit.jsf?name=#{pageSession.encodedName}&isAppScopedRes=$boolean{true}&appName=#{pageSession.encodedAppName}&moduleName=#{pageSession.encodedModuleName}");
             setPageSessionAttribute(key="parentUrl", value="#{sessionScope.REST_URL}/applications/application/#{pageSession.encodedAppName}/module/#{pageSession.encodedModuleName}/resources");
         }
         if(!#{pageSession.moduleName}) {
-            setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/full/mailResourceEdit.jsf?name=#{pageSession.encodedName}&isAppScopedRes=$boolean{true}&appName=#{pageSession.encodedAppName}");
+            setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/jmail/mailResourceEdit.jsf?name=#{pageSession.encodedName}&isAppScopedRes=$boolean{true}&appName=#{pageSession.encodedAppName}");
             setPageSessionAttribute(key="parentUrl", value="#{sessionScope.REST_URL}/applications/application/#{pageSession.encodedAppName}/resources");
         }
         setPageSessionAttribute(key="parentPage" value="#{request.contextPath}/common/applications/resources/appScopedResources.jsf?appName=#{pageSession.encodedAppName}&listPageLink=/common/applications/applications.jsf");
     }
     if(!#{pageSession.isAppScopedRes}){
-        setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/full/mailResourceEdit.jsf?name=#{pageSession.encodedName}");
-        setPageSessionAttribute(key="parentPage" value="#{request.contextPath}/full/mailResources.jsf");
+        setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/jmail/mailResourceEdit.jsf?name=#{pageSession.encodedName}");
+        setPageSessionAttribute(key="parentPage" value="#{request.contextPath}/jmail/jakartaMailResources.jsf");
         setPageSessionAttribute(key="parentUrl", value="#{sessionScope.REST_URL}/resources");
     }
     setPageSessionAttribute(key="childType" value="mail-resource");
@@ -79,7 +79,7 @@
 #include "/common/resourceNode/resourceEditPageButtons.inc"
     </sun:title>
 
-#include "/full/mailResourceAttr.inc"
+#include "/jmail/mailResourceAttr.inc"
 #include "/common/shared/propertyDescTable.inc"
     "<br/><br/>
