+++ appserver/admingui/mail-plugin/src/main/resources/mailResourceNew.jsf
@@ -16,19 +16,19 @@
 
 -->
 
-<!-- full/mailResourceNew.jsf -->
+<!-- jmail/mailResourceNew.jsf -->
 
 <!initPage
-    setResourceBundle(key="i18njmail" bundle="org.glassfish.full.admingui.Strings");
-    setResourceBundle(key="help_jmail" bundle="org.glassfish.full.admingui.Helplinks")
+    setResourceBundle(key="i18njmail" bundle="org.glassfish.jmail.admingui.Strings");
+    setResourceBundle(key="help_jmail" bundle="org.glassfish.jmail.admingui.Helplinks")
 />
 <!composition template="/templates/default.layout"  guiTitle="$resource{i18njmail.javaMail.newPageTitle}" >
 <!define name="content">
     <event>
     <!beforeCreate
         getRequestValue(key="target" value="#{pageSession.targetValue}");
-        setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/full/mailResourceNew.jsf");
-        setPageSessionAttribute(key="parentPage" value="#{request.contextPath}/full/mailResources.jsf");
+        setPageSessionAttribute(key="selfPage" value="#{request.contextPath}/jmail/mailResourceNew.jsf");
+        setPageSessionAttribute(key="parentPage" value="#{request.contextPath}/jmail/jakartaMailResources.jsf");
         setPageSessionAttribute(key="childType" value="mail-resource");
         
         setPageSessionAttribute(key="parentUrl", value="#{sessionScope.REST_URL}/resources");
@@ -56,7 +56,7 @@
     <sun:title id="propertyContentPage" title="$resource{i18njmail.javaMail.newPageTitle}" helpText="$resource{i18njmail.javaMail.newPageTitleHelp}">
 #include "/common/resourceNode/resourceEditPageButtons.inc"
     </sun:title>
-#include "/full/mailResourceAttr.inc"
+#include "/jmail/mailResourceAttr.inc"
 #include "/common/shared/propertyDescTable.inc"
 <event>
 <!afterCreate
