+++ appserver/admingui/mail-plugin/src/main/resources/org/glassfish/jmail/admingui/Strings.properties
@@ -1,0 +1,52 @@
+#
+# Copyright (c) 2009, 2020 Oracle and/or its affiliates. All rights reserved.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License v. 2.0, which is available at
+# http://www.eclipse.org/legal/epl-2.0.
+#
+# This Source Code may also be made available under the following Secondary
+# Licenses when the conditions for such availability set forth in the
+# Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+# version 2 with the GNU Classpath Exception, which is available at
+# https://www.gnu.org/software/classpath/license.html.
+#
+# SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+#
+
+resourcesTable.sessions=Sessions
+resourcesTable.colJndiName=JNDI Name
+resourcesTable.resources=Resources
+resourcesTable.colResType=Resource Type
+resourcesTable.colDesc=Description
+
+tree.javaMailSessions=Jakarta Mail Sessions
+tree.javaMailSessions.tooltip=Jakarta Mail Sessions
+
+## Jakarta Mail Sessions
+javaMailSessions.pageTitle=Jakarta Mail Sessions
+javaMailSessions.pageTitleHelp=A Jakarta Mail session resource represents a mail session in the Jakarta Mail API, which provides a platform-independent and protocol-independent framework to build mail and messaging applications.
+javaMail.editPageTitle=Edit Jakarta Mail Session
+javaMail.editPageTitleHelp=A Jakarta Mail session resource represents a mail session in the Jakarta Mail API.
+javaMail.newPageTitle=New Jakarta Mail Session
+javaMail.newPageTitleHelp=A Jakarta Mail session resource represents a mail session in the Jakarta Mail API.
+javaMail.targetPageTitle=Jakarta Mail Session Targets
+javaMail.jndiHelp=A unique identifier; must contain only alphanumeric, underscore, dash, or dot characters
+javaMail.mailHost=Mail Host:
+javaMail.mailHostHelp=DNS name of the default mail server
+javaMail.user=Default User:
+javaMail.userHelp=User name to provide when connecting to a mail server; must contain only alphanumeric, underscore, dash, or dot characters
+javaMail.returnAddr=Default Sender Address:
+javaMail.returnAddrHelp=E-mail address of the default user
+javaMail.descHelp=Makes it easier to find this session later
+javaMail.Session=Session:
+javaMail.storeProtocol=Store Protocol:
+javaMail.storeProtocolHelp=Either IMAP or POP3; default is IMAP
+javaMail.storeClass=Store Protocol Class:
+javaMail.storeClassHelp=Default is com.sun.mail.imap.IMAPStore
+javaMail.transportProtocol=Transport Protocol:
+javaMail.transportProtocolHelp=Default is SMTP
+javaMail.transProtocolClass=Transport Protocol Class:
+javaMail.transProtocolClassHelp=Default is com.sun.mail.smtp.SMTPTransport
+javaMail.Debug=Debug:
+javaMail.DebugHelp=Select the Debug checkbox to enable extra debugging output for this mail session, including a protocol trace.
