+++ appserver/featuresets/glassfish/pom.xml
@@ -773,6 +773,17 @@
         </dependency>
         <dependency>
             <groupId>org.glassfish.main.admingui</groupId>
+            <artifactId>console-mail-plugin</artifactId>
+            <version>${project.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>*</groupId>
+                    <artifactId>*</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.glassfish.main.admingui</groupId>
             <artifactId>console-jms-plugin</artifactId>
             <version>${project.version}</version>
             <exclusions>
